Update powerRule.py
Added dots, with one dot moving to the other dot. Most of the dot code was made using ChatGPT since I'm not very familiar with manim (like the update from alpha function and the update dot code). I took the code it generated and modified it to my own likings.