Attempt to write slope
Something with the slope formula is wrong and it displays in the middle of the screen, but it's still progress. Next steps will involve positioning, rounding to avoid floating point error, and stopping the NAN from coming out, probably by making the other point stop epsilon away from the other point.