Directory Structure Update
Restructured the directory so that all files used in the actual running
of opticos are stored in a directory called 'source'. Additionally added
the introduction to the Opticos user manual which is stored in
'user_manual' in the root of the repository. Additionally put the extras
directory outside the source directory. Finally, the read me files were
changed from .txt files to .md files.